data_micro/requirements.txt
environs==14.1.1
marshmallow==4.0.0
nats-py==2.10.0
python-dotenv==1.1.0
